+++ src/_pages/SpecializationPage/ui/SpecializationPage.tsx
@@ -1,6 +1,5 @@
 import { Collection } from '@/entities/collection';
 import { HhTopBySpecResponse } from '@/entities/hh';
-import { getQuestionsList } from '@/entities/question';
 import { Specialization } from '@/entities/specialization';
 import { Flex } from '@/shared/ui/Flex';
 import { CollectionsSection } from '@/widgets/Specialization/CollectionsSection';
@@ -8,7 +7,7 @@ import { Header } from '@/widgets/Specialization/Header';
 import { KeywordsSection } from '@/widgets/Specialization/KeywordsSection';
 import { SkillsSection } from '@/widgets/Specialization/SkillsSection';
 import { TasksSection } from '@/widgets/Specialization/TasksSection';
-import { QuestionsBlock } from '@/widgets/question/SpecializationsQuestionBlock';
+import { QuestionsSection } from '@/widgets/question/QuestionsSection';
 
 import styles from './SpecializationPage.module.css';
 
@@ -18,30 +17,19 @@ interface SpecializationPageProps {
 	specAnalytics?: HhTopBySpecResponse;
 	collections: Collection[];
 }
-const FIRST_PAGE = 1;
-const MAX_QUSTIONS_COUNT = 4;
 
-export const SpecializationPage = async ({
+export const SpecializationPage = ({
 	specialization,
 	specAnalytics,
 	collections,
 	locale,
 }: SpecializationPageProps) => {
-	const questionsResponse = await getQuestionsList({
-		page: FIRST_PAGE,
-		limit: MAX_QUSTIONS_COUNT,
-		specializationId: specialization.id,
-	});
 	return (
 		<Flex direction="column" gap="40" className={styles.container}>
 			<Header specialization={specialization} />
 			<SkillsSection skills={specAnalytics?.skills} />
 			<KeywordsSection keywords={specAnalytics?.keywords} />
-			<QuestionsBlock
-				locale={locale}
-				specialization={specialization.slug}
-				questions={questionsResponse.data}
-			/>
+			<QuestionsSection locale={locale} specialization={specialization} />
 			<CollectionsSection
 				collections={collections}
 				specializationSlug={specialization.slug}
