+++ src/entities/company/ui/CompanyCompactList/CompanyCompactList.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 
-import EmptyStub from '@/shared/assets/icons/EmptyStub.svg';
+import EmptyStub from '@/shared/assets/icons/emptyStub.svg';
 import { Flex } from '@/shared/ui/Flex';
 import { Tooltip } from '@/shared/ui/Tooltip';
