+++ src/widgets/question/QuestionsSection/index.ts
@@ -1,0 +1,1 @@
+export { QuestionsSection } from './ui/QuestionsSection/QuestionsSection';
