--- src/widgets/question/SpecializationsQuestionBlock/index.ts
@@ -1,3 +1,0 @@
-export { QuestionsBlock } from './ui/QuestionsBlock/QuestionsBlock';
-export { InterviewQuestionsList } from './ui/InterviewQuestionsList/InterviewQuestionsList';
-export { InterviewQuestionsItem } from './ui/InterviewQuestionsItem/InterviewQuestionsItem';
